moved some Transaction logic to TransactionRepository
